backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
pgvector>=0.3.6
pydantic-settings>=2.6.0
python-multipart>=0.0.12
google-genai>=1.0.0
google-cloud-texttospeech>=2.21.0
alembic>=1.14.0
httpx>=0.28.0
google-cloud-speech>=2.27.0
